refactor Event Listener to CustomEventListener for clean up code and single responsibility
